mount: merge aufs and overlay mount types
The new 'overlay' mount type requires 3 arguments like the old, even if AuFS
is used. Also, the check for AuFS or OverlayFS is currently done at build-time,
but at some point in the future I'd like to do it at runtime somehow.

The aufs type is still available for backwards compatibility, but is not
documented.