pflask: properly create and configure /dev/console
mknod() is not allowed inside unprivileged user namespaces, so we only
bind mount it.

Additionally, we properly change /dev/console owner before creating the
container. This is needed because for an unprivileged user with uid/gid
X to chown a file to subuid/subgid Y, he needs to run chown as root in
a user namespace where nsid 0 is mappend to hostuid/hostgid Y, and nsid
Y is mapped to hostuid/hostgid X, so that the container root is privileged
with respect to hostuid/hostgid X, allowing him to do the chown.